move main out of __init__ into its own file
